# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"errors": {},
					"path": {
						"type": "string"
					},
					"timestamp": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"path",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnitTypeStatusEnum": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
			"UnitTypeCategoryEnum": {
				"enum": [
					"locker",
					"small",
					"medium",
					"large",
					"climate_controlled"
				],
				"type": "string"
			},
...
			"ApiResponse_PaginatedData_UnitTypeResponse__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginatedData_UnitTypeResponse_"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_UnitTypeResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/UnitTypeResponse"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateUnitTypeRequest": {
				"$ref": "#/components/schemas/infer_typeofcreateUnitTypeSchema_"
			},
...
			"UpdateUnitTypeRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateUnitTypeSchema_"
			},
...
			"BillingUnitEnum": {
				"enum": [
					"day",
					"month",
					"one_time"
				],
				"type": "string"
			},
			"FacilityUnitTypeOfferingStatusEnum": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
...
			"ApiResponse_PaginatedData_FacilityUnitTypeOfferingResponse__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginatedData_FacilityUnitTypeOfferingResponse_"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter your JWT access token"
			},
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "accessToken",
				"description": "Access token stored in HttpOnly cookie"
			}
		}
	},
	"info": {
		"title": "project_be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Van Thai"
		}
	},
	"paths": {
...
		"/unit-types": {
			"get": {
				"operationId": "FindAllUnitType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedData_UnitTypeResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UnitTypeCategoryEnum"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UnitTypeStatusEnum"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"name",
								"createdAt",
								"updatedAt",
								"area",
								"volume"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUnitType",
				"responses": {
					"201": {
						"description": "Unit type created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UnitTypeResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or unit type already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUnitTypeRequest"
							}
						}
					}
				}
			}
		},
		"/unit-types/{id}": {
			"get": {
				"operationId": "FindUnitTypeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UnitTypeResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Unit type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUnitType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UnitTypeResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Unit type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUnitTypeRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUnitType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Unit type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/unit-types/{id}/restore": {
			"post": {
				"operationId": "RestoreUnitType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UnitTypeResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Unit type is not deleted or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Unit type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Unit Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/facility-unit-type-offerings": {
			"get": {
				"operationId": "GetOfferings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedData_FacilityUnitTypeOfferingResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get list of facility unit type offerings (with pagination, filter, sorting)",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/FacilityUnitTypeOfferingStatusEnum"
						}
					},
					{
						"in": "query",
						"name": "billingUnit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BillingUnitEnum"
						}
					},
					{
						"in": "query",
						"name": "unitTypeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "facilityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"createdAt",
								"updatedAt",
								"pricePerUnit",
								"minRentalDays"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOffering",
				"responses": {
					"201": {
						"description": "Offering created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error, inactive Facility/UnitType, or offering already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility or Unit Type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOfferingRequest"
							}
						}
					}
				}
			}
		},
		"/facility-unit-type-offerings/{id}": {
			"get": {
				"operationId": "GetOfferingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get facility unit type offering details by ID",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOfferingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Soft-delete facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-unit-type-offerings/facility/{facilityId}/unit-type/{unitTypeId}": {
			"get": {
				"operationId": "GetOfferingByFacilityAndUnitType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Quick lookup offering by facilityId and unitTypeId\r\n(used by StorageUnit, Reservation, Contract)",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "facilityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "unitTypeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-unit-type-offerings/{id}/restore": {
			"post": {
				"operationId": "RestoreOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Offering is not deleted or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings": {
			"get": {
				"operationId": "GetOfferings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedData_FacilityAmenityOfferingResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/FacilityAmenityOfferingStatusEnum"
						}
					},
					{
						"in": "query",
						"name": "billingUnit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BillingUnitEnum"
						}
					},
					{
						"in": "query",
						"name": "amenityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "facilityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"status",
								"createdAt",
								"updatedAt",
								"pricePerUnit",
								"totalQuantity",
								"inUseQuantity"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOffering",
				"responses": {
					"201": {
						"description": "Facility amenity offering created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or offering already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateFacilityAmenityOfferingRequest"
							}
						}
					}
				}
			}
		},
...
		"/facility-amenity-offerings/{id}": {
			"get": {
				"operationId": "GetOfferingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateFacilityAmenityOfferingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format or items in use",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings/{id}/restore": {
			"post": {
				"operationId": "RestoreOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Offering is not deleted or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	]
}